+++ PuppyWorld/src/main/java/com/example/ad/domain/Services.java
@@ -11,7 +11,7 @@ public class Services {
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private long serviceId;
+	private int serviceId;
 	
 	private String serviceName, serviceDescription, serviceDuration, charges;
 	
@@ -34,11 +34,11 @@ public class Services {
 		this.limitDogs = limitDogs;
 	}
 
-	public long getServiceId() {
+	public int getServiceId() {
 		return serviceId;
 	}
 
-	public void setServiceId(long serviceId) {
+	public void setServiceId(int serviceId) {
 		this.serviceId = serviceId;
 	}
 
@@ -97,5 +97,6 @@ public class Services {
 	public void setLimitDogs(LimitDogs limitDogs) {
 		this.limitDogs = limitDogs;
 	}
+
 
 }
